requirements.txt
langchain
python-dotenv
groq
langchain-groq
streamlit
-e .